Change FileSystem wrapper from additional syntax to a facade
